move check_exchange to exchange package
