+++ freqtrade/exchange/check_exchange.py
